+++ Editor/Enums/EFieldType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Unity_Free_Online_Config.Editor.Enums
 {
-    public enum EPropertyType
+    public enum EFieldType
     {
         String,
         Int,