+++ Editor/Scripts/CreateNewConfigView.cs
@@ -1,11 +1,13 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Text;
+using Unity_Free_Online_Config.Editor.Enums;
 using UnityEditor;
 using UnityEngine;
 
 namespace Unity_Free_Online_Config.Editor.Scripts
 {
-
     public class CreateNewConfigView : EditorWindow
     {
         private string _newConfigName;
@@ -14,22 +16,25 @@ namespace Unity_Free_Online_Config.Editor.Scripts
         private bool _embedNameIntoFile;
         private bool _embedUrlIntoFile;
         private Texture2D _backgroundTexture;
+        private float baseTextureCoordsSize;
         private Rect _backgroundTextureCoords;
         private Vector2 _windowMinSizeVector;
         private Vector2 _windowMaxSizeVector;
-        private const string MenuItem = "VOUFO/Create New Config";
-        private const string WindowLabel = "Create new Spreadsheet config";
-        private const string ConfigNameLabel = "Config Name";
-        private const string SpreadsheetURLLabel = "Spreadsheet URL";
-        private const string PathURLLabel = "Path for generated files";
-        private const string DefaultPath = "Assets/VUFOC/<ConfigName>/";
-        
+        private ConfigVoFieldsScriptableObject _fieldsHolder;
+        private SerializedObject _fieldsSerializedObject;
+        private SerializedProperty _fieldsSerializedProperty;
         private GUILayoutOption[] _headerLabelOptions;
         private GUIStyle _headerLabelStyle;
         private GUILayoutOption[] _defaultLabelOptions;
         private GUIStyle _defaultLabelStyle;
         private GUILayoutOption[] _buttonOptions;
         private GUIStyle _buttonStyle;
+        private const string MenuItem = "VOUFO/Create New Config";
+        private const string WindowLabel = "Create new Spreadsheet config";
+        private const string ConfigNameLabel = "Config Name";
+        private const string SpreadsheetURLLabel = "Spreadsheet URL";
+        private const string PathURLLabel = "Path for generated files";
+        private const string DefaultPath = "Assets/VUFOC/<ConfigName>/";
         
         [MenuItem (MenuItem)]
         public static void  ShowWindow () => 
@@ -92,52 +97,73 @@ namespace Unity_Free_Online_Config.Editor.Scripts
             }
             
             _backgroundTexture.Apply();
-            _backgroundTextureCoords = new Rect(0, 0, 10, 10);
+            baseTextureCoordsSize = 10;
+            _backgroundTextureCoords = new Rect(0, 0, baseTextureCoordsSize, baseTextureCoordsSize);
             _windowMinSizeVector = new Vector2(500, 500);
-            _windowMaxSizeVector = new Vector2(500, 500);
+            _windowMaxSizeVector = new Vector2(500, 3000);
             
             minSize = _windowMinSizeVector;
             maxSize = _windowMaxSizeVector;
+            
+            _fieldsHolder = CreateInstance<ConfigVoFieldsScriptableObject>();
+            _fieldsHolder.ConfigVoFields = new List<ConfigVoField>();
+            _fieldsSerializedObject = new SerializedObject(_fieldsHolder);
+            _fieldsSerializedProperty = _fieldsSerializedObject.FindProperty("ConfigVoFields");
         }
-
+        
         private void OnGUI()
         {
+            var ratio = rootVisualElement.contentRect.width / rootVisualElement.contentRect.height;
+            var backgroundCoordHeight = baseTextureCoordsSize / ratio;
+            
             if(_defaultLabelOptions == null || 
                _defaultLabelStyle == null ||
-               _backgroundTexture == null)
+               _backgroundTexture == null ||
+               _fieldsSerializedObject == null)
                 InitializeGUIResources();
 
+            EditorGUILayout.Space(10f);
             EditorGUILayout.BeginHorizontal();
             EditorGUILayout.Space(10f);
             EditorGUILayout.BeginVertical();
             
             // Background
+            _backgroundTextureCoords = new Rect(0, -backgroundCoordHeight, baseTextureCoordsSize, backgroundCoordHeight);
             GUI.DrawTextureWithTexCoords(rootVisualElement.contentRect, _backgroundTexture, _backgroundTextureCoords);
             
             // Name
-            EditorGUILayout.LabelField(WindowLabel, _headerLabelStyle, _headerLabelOptions);
+            //EditorGUILayout.LabelField(WindowLabel, _headerLabelStyle, _headerLabelOptions);
         
+            // Path for generated databases
+            EditorGUILayout.BeginHorizontal();
+            EditorGUILayout.LabelField(PathURLLabel, _defaultLabelStyle, _defaultLabelOptions);
+            GUILayout.Box($"Using default path: {_targetPath}", GUILayout.ExpandWidth(true));
+            EditorGUILayout.EndHorizontal();
+
+            // Embed Name Toggle
+            _embedNameIntoFile = EditorGUILayout.ToggleLeft("Embed name into file", _embedNameIntoFile, _defaultLabelStyle, _defaultLabelOptions);
+            
             // Config Name input
             EditorGUILayout.BeginHorizontal();
             EditorGUILayout.LabelField(ConfigNameLabel, _defaultLabelStyle, _defaultLabelOptions);
             _newConfigName = EditorGUILayout.TextField(_newConfigName);
             EditorGUILayout.EndHorizontal();
-        
-            // Config Url input
-            EditorGUILayout.BeginHorizontal();
-            EditorGUILayout.LabelField(SpreadsheetURLLabel, _defaultLabelStyle, _defaultLabelOptions);
-            _newConfigUrl = EditorGUILayout.TextField(_newConfigUrl);
-            EditorGUILayout.EndHorizontal();
             
-            // Path for generated databases
-            EditorGUILayout.BeginHorizontal();
-            EditorGUILayout.LabelField(PathURLLabel, _defaultLabelStyle, _defaultLabelOptions);
-            //_targetPath = EditorGUILayout.TextField(_targetPath);
-            GUILayout.Box($"Using default path: {_targetPath}", GUILayout.ExpandWidth(true));
-            EditorGUILayout.EndHorizontal();
-
-            _embedNameIntoFile = EditorGUILayout.ToggleLeft("Embed name into file", _embedNameIntoFile, _defaultLabelStyle, _defaultLabelOptions);
+            // Embed URL Toggle
             _embedUrlIntoFile = EditorGUILayout.ToggleLeft("Embed URL into file", _embedUrlIntoFile, _defaultLabelStyle, _defaultLabelOptions);
+
+            
+            if (_embedUrlIntoFile)
+            {
+                // Config Url input
+                EditorGUILayout.BeginHorizontal();
+                EditorGUILayout.LabelField(SpreadsheetURLLabel, _defaultLabelStyle, _defaultLabelOptions);
+                _newConfigUrl = EditorGUILayout.TextField(_newConfigUrl);
+                EditorGUILayout.EndHorizontal();
+            }
+            
+            // List of fields in VO
+            EditorGUILayout.PropertyField(_fieldsSerializedProperty);
             
             //Create button
             EditorGUILayout.BeginHorizontal();
@@ -153,11 +179,42 @@ namespace Unity_Free_Online_Config.Editor.Scripts
             EditorGUILayout.Space(10f);
             EditorGUILayout.EndHorizontal();
         }
-
+        
         private void OnCreateInput()
         {
             var fullPathConfig = _targetPath + $"{_newConfigName}/{_newConfigName}SpreadsheetConfig.cs";
             var fullPathVo = _targetPath + $"{_newConfigName}/{_newConfigName}SpreadsheetVo.cs";
+            var propEnum = _fieldsSerializedProperty.GetEnumerator();
+            var configVoFields = _fieldsSerializedProperty;
+            var stringBuilder = new StringBuilder();
+            propEnum.MoveNext();
+            
+            for (var i = 0; i < configVoFields.arraySize; ++i)
+            {
+                var voField = (ConfigVoField)((SerializedProperty)propEnum.Current)?.boxedValue;
+                var fieldType = voField.FieldType;
+                var fieldName = voField.FieldName;
+                var publicPrefix = "public ";
+
+                string type = fieldType switch
+                {
+                    EFieldType.String => "string",
+                    EFieldType.Int => "int",
+                    EFieldType.Float => "float",
+                    EFieldType.Double => "double",
+                    _ => ""
+                };
+            
+                stringBuilder.Append("\t");
+                stringBuilder.Append(publicPrefix);
+                stringBuilder.Append(type);
+                stringBuilder.Append(" ");
+                stringBuilder.Append(fieldName);
+                stringBuilder.Append(";\n");
+                propEnum.MoveNext();
+            }
+            
+            (propEnum as IDisposable).Dispose();
             
             Directory.CreateDirectory(_targetPath + _newConfigName);
             File.Create(fullPathConfig).Close();
@@ -165,37 +222,38 @@ namespace Unity_Free_Online_Config.Editor.Scripts
             
             File.WriteAllLines(fullPathVo, new string[]
             {
-$@"
-using System;
+$@"using System;
 using Models;
 // Auto-generated file
 [Serializable]
 public class {_newConfigName}SpreadsheetVo : ASpreadsheetVo
 {{
-    public string ExampleString;
-    public int ExampleInt;
-    public float ExampleFloat;
-    public double ExampleDouble;
-}}
-"
+{stringBuilder}
+}}"
             });
+
+            var urlProperty = _embedUrlIntoFile
+                ? $"\tpublic override string DatabaseDataUrl => \"{_newConfigUrl}\";"
+                : "\tpublic override string DatabaseDataUrl => databaseDataUrl;";
+
+            var urlField = _embedUrlIntoFile
+                ? ""
+                : "\n\t[SerializeField] private string databaseDataUrl;";
             
             File.WriteAllLines(fullPathConfig, new string[]
             {
-$@"
-using UnityEngine;
+$@"using UnityEngine;
 using Models;
 // Auto-generated file
 [CreateAssetMenu(fileName = ""{_newConfigName}"", menuName = ""Databases/{_newConfigName}"")]
 public class {_newConfigName}SpreadsheetDatabase : ASpreadsheetDatabase<{_newConfigName}SpreadsheetVo>
 {{
-    [SerializeField] private string databaseName;
-    [SerializeField] private string databaseDataUrl;
+    [SerializeField] private string databaseName;{urlField}
+    
     public override string DatabaseName => databaseName;
-    public override string DatabaseDataUrl => databaseDataUrl;
+    {urlProperty}
     protected override ASpreadsheetVo TemplateVo => new {_newConfigName}SpreadsheetVo();
-}}
-"
+}}"
             });
             
             AssetDatabase.Refresh();